+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Fixes
 
+- Persistent stack overflow fix for Windows build in [#365](https://github.com/TNO-S3/WuppieFuzz/pull/365)
+
 # v1.7.0 (2026-08-26)
 
 ## Highlights